+++ src/Rule/CharacterClass.php
@@ -8,6 +8,7 @@ use InvalidArgumentException;
 use StableSort\StableSort;
 use Stadly\PasswordPolice\Constraint\Count;
 use Stadly\PasswordPolice\Password;
+use Stadly\PasswordPolice\ValidationError;
 
 abstract class CharacterClass implements RuleInterface
 {
@@ -78,19 +79,21 @@ abstract class CharacterClass implements RuleInterface
     }
 
     /**
-     * Enforce that a password is in compliance with the rule.
+     * Validate that a password is in compliance with the rule.
      *
-     * @param Password|string $password Password that must adhere to the rule.
-     * @throws RuleException If the password does not adhrere to the rule.
+     * @param Password|string $password Password to validate.
+     * @return ValidationError|null Validation error describing why the password is not in compliance with the rule.
      */
-    public function enforce($password): void
+    public function validate($password): ?ValidationError
     {
         $count = $this->getCount((string)$password);
         $constraint = $this->getViolation($count);
 
         if ($constraint !== null) {
-            throw new RuleException($this, $constraint->getWeight(), $this->getMessage($constraint, $count));
+            return new ValidationError($this, $constraint->getWeight(), $this->getMessage($constraint, $count));
         }
+
+        return null;
     }
 
     /**