+++ tests/Rule/LowerCaseTest.php
@@ -6,6 +6,7 @@ namespace Stadly\PasswordPolice\Rule;
 
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
+use Stadly\PasswordPolice\ValidationError;
 
 /**
  * @coversDefaultClass \Stadly\PasswordPolice\Rule\LowerCase
@@ -192,88 +193,77 @@ final class LowerCaseTest extends TestCase
     }
 
     /**
-     * @covers ::enforce
+     * @covers ::validate
      */
-    public function testEnforceDoesNotThrowExceptionWhenRuleIsSatisfied(): void
+    public function testRuleCanBeValidated(): void
     {
         $rule = new LowerCase(1, null);
 
-        $rule->enforce('foo');
-
-        // Force generation of code coverage
-        $ruleConstruct = new LowerCase(1, null);
-        self::assertEquals($rule, $ruleConstruct);
+        self::assertNull($rule->validate('foo'));
     }
 
     /**
-     * @covers ::enforce
+     * @covers ::validate
      */
-    public function testEnforceThrowsExceptionWhenRuleIsNotSatisfied(): void
-    {
-        $rule = new LowerCase(1, null);
-
-        $this->expectException(RuleException::class);
-
-        $rule->enforce('FOO');
-    }
-
-    /**
-     * @covers ::enforce
-     */
-    public function testValidationMessageForRuleWithMinConstraint(): void
+    public function testRuleWithMinConstraintCanBeInvalidated(): void
     {
         $rule = new LowerCase(5, null);
 
-        $this->expectExceptionMessage('There must be at least 5 lower case characters.');
-
-        $rule->enforce('Foo');
+        self::assertEquals(
+            new ValidationError($rule, 1, 'There must be at least 5 lower case characters.'),
+            $rule->validate('Foo')
+        );
     }
 
     /**
-     * @covers ::enforce
+     * @covers ::validate
      */
-    public function testValidationMessageForRuleWithMaxConstraint(): void
+    public function testRuleWithMaxConstraintCanBeInvalidated(): void
     {
         $rule = new LowerCase(0, 10);
 
-        $this->expectExceptionMessage('There must be at most 10 lower case characters.');
-
-        $rule->enforce('Foo bar qwerty test');
+        self::assertEquals(
+            new ValidationError($rule, 1, 'There must be at most 10 lower case characters.'),
+            $rule->validate('Foo bar qwerty test')
+        );
     }
 
     /**
-     * @covers ::enforce
+     * @covers ::validate
      */
-    public function testValidationMessageForRuleWithBothMinAndMaxConstraint(): void
+    public function testRuleWithBothMinAndMaxConstraintCanBeInvalidated(): void
     {
         $rule = new LowerCase(5, 10);
 
-        $this->expectExceptionMessage('There must be between 5 and 10 lower case characters.');
-
-        $rule->enforce('Foo');
+        self::assertEquals(
+            new ValidationError($rule, 1, 'There must be between 5 and 10 lower case characters.'),
+            $rule->validate('Foo')
+        );
     }
 
     /**
-     * @covers ::enforce
+     * @covers ::validate
      */
-    public function testValidationMessageForRuleWithMaxConstraintEqualToZero(): void
+    public function testRuleWithMaxConstraintEqualToZeroCanBeInvalidated(): void
     {
         $rule = new LowerCase(0, 0);
 
-        $this->expectExceptionMessage('There must be no lower case characters.');
-
-        $rule->enforce('Foo');
+        self::assertEquals(
+            new ValidationError($rule, 1, 'There must be no lower case characters.'),
+            $rule->validate('Foo')
+        );
     }
 
     /**
-     * @covers ::enforce
+     * @covers ::validate
      */
-    public function testValidationMessageForRuleWithMinConstraintEqualToMaxConstraint(): void
+    public function testRuleWithMinConstraintEqualToMaxConstraintCanBeInvalidated(): void
     {
         $rule = new LowerCase(3, 3);
 
-        $this->expectExceptionMessage('There must be exactly 3 lower case characters.');
-
-        $rule->enforce('Foo');
+        self::assertEquals(
+            new ValidationError($rule, 1, 'There must be exactly 3 lower case characters.'),
+            $rule->validate('Foo')
+        );
     }
 }